requirements.txt
streamlit
Pillow
torch
torchvision
scikit-learn
matplotlib
numpy
pandas
